(!! break) Move all is* function to `assert` and rename `assert` to `is`
It was strange to have to import predicates from all over the place so
they all got moved to a single package: `assert`.

The name assert was kind of misleading. "assert" sounds like something that
stops the program when a condition is not met, like the node's own assert package.
`is` sounds good to me. Now I can do:

```
const is = require('tolb/is');
```
or
```
import * as is from 'tolb/next/is';
```

and code such as:

```
if (!is.string(foo)) {
  ...
}

list.reject(is.empty, bar);
```

In general, functions got new names by just removing the `is` from their names
but there are some exceptions:

* `isFunction`, (from `object`) couldn't be renamed to `function` because function
   is a reserved word, so `object.isFunction` became `is.func`
* `isNaN`, (from `number`), if renamed to `NaN`, would conflict with the NaN
  object, so now `number.isNaN` is `is.nan`
* `equals` (from `assert`) would sound very weird is kept on the plural form so
  now, `assert.equals` is `is.equal`

All functions in `number` were moved, so that package no longer exists